Refactor project and Fix API

+++ TodoList.Classes/TodoRepository.cs
@@ -1,50 +1,57 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Text.Json;
-namespace TodoList.Classes{
-   public class TodoRepository
-   {
+using System.Threading.Tasks;
 
-    private const string BasePathDirectory=$"./taskUser/";
-    
-    private string UserPathDirectory{get;set;}
-
-    public string nameUser {get;set;}
-    
-    public TodoRepository(string nickname)
+namespace TodoList.Classes
+{
+    public static class TodoRepository
     {
-        this.nameUser=nickname;
-        this.UserPathDirectory=Path.Combine(BasePathDirectory,$"{nickname}/");
-    }
-        public bool Add(UserTask task){
-            var id = Guid.NewGuid();
-            task.Id=id;
-            var json = JsonSerializer.Serialize(task,new JsonSerializerOptions{WriteIndented=true});
-            var filename=$"{id}.json";
-            var fullpath= Path.Combine(UserPathDirectory,filename);
-            if(!Directory.Exists(UserPathDirectory))
+        private static string BasePathDirectory = $"./taskUser/";
+        private static string BaseDirectory = AppDomain.CurrentDomain.BaseDirectory;
+
+        private static string FullDirectory { get; set; }
+
+        public static string UserName { get; set; }
+
+
+        public static bool Enter(string userName)
+        {
+
+            UserName = userName;
+            FullDirectory = Path.Combine(BaseDirectory, BasePathDirectory, UserName);
+
+            if (!Directory.Exists(FullDirectory))
             {
-                Directory.CreateDirectory(UserPathDirectory);
+                Directory.CreateDirectory(FullDirectory);
             }
-            File.WriteAllText(fullpath,json);
+
             return true;
         }
 
-        public UserTask[] Get()
+        public static bool Add(UserTask task)
         {
-            if(!Directory.Exists(UserPathDirectory))
-            {
-                Directory.CreateDirectory(UserPathDirectory);
-            }
-            var files=Directory.GetFiles(UserPathDirectory);
+            var json = JsonSerializer.Serialize(task, new JsonSerializerOptions { WriteIndented = true });
+            var filename = $"{task.Id}.json";
+            var fullpath = Path.Combine(FullDirectory, filename);
+            File.WriteAllText(fullpath, json);
+            return true;
+        }
+
+        public static UserTask[] Get()
+        {
+            var files = Directory.GetFiles(FullDirectory);
             var tasks = new List<UserTask>();
             foreach (var item in files)
             {
                 var json = File.ReadAllText(item);
                 var task = JsonSerializer.Deserialize<UserTask>(json);
-                if(task!=null)
+                if (task != null)
                 {
                     tasks.Add(task);
-                }else
+                }
+                else
                 {
                     throw new Exception("Task cannot deserialized");
                 }
@@ -52,21 +59,61 @@ namespace TodoList.Classes{
             return tasks.ToArray();
         }
 
-        public void Edit(UserTask task){
-           var fullpath = $"{UserPathDirectory}/{task.Id}.json";
-           var item = File.ReadAllText(fullpath);
-           var jsonObj = JsonSerializer.Deserialize<UserTask>(item);
-           jsonObj.IsClose=task.IsClose;
-           jsonObj.Note=task.Note;
-           var output = JsonSerializer.Serialize(jsonObj,new JsonSerializerOptions{WriteIndented=true});
-           File.WriteAllText(fullpath,output);
+        public static bool Edit(Guid id, string note)
+        {
+            var fullpath = $"{FullDirectory}/{id}.json";
+
+            if (!File.Exists(fullpath))
+            {
+                return false;
+            }
+
+            var item = File.ReadAllText(fullpath);
+            var jsonObj = JsonSerializer.Deserialize<UserTask>(item);
+            if (note != null)
+            {
+                jsonObj.Note = note;
+            }
+
+            var output = JsonSerializer.Serialize(jsonObj, new JsonSerializerOptions { WriteIndented = true });
+            File.WriteAllText(fullpath, output);
+            return true;
         }
 
-        public void Delete(UserTask task){
-            var fullpath = $"{UserPathDirectory}/{task.Id}.json";
+        public static bool Edit(Guid id, bool close)
+        {
+            var fullpath = $"{FullDirectory}/{id}.json";
+
+            if (!File.Exists(fullpath))
+            {
+                return false;
+            }
+
+            var item = File.ReadAllText(fullpath);
+            var jsonObj = JsonSerializer.Deserialize<UserTask>(item);
+
+            jsonObj.IsClose = close;
+
+
+            var output = JsonSerializer.Serialize(jsonObj, new JsonSerializerOptions { WriteIndented = true });
+            File.WriteAllText(fullpath, output);
+            return true;
+        }
+
+
+
+        public static bool Delete(Guid id)
+        {
+            var fullpath = $"{FullDirectory}/{id}.json";
+
+            if (!File.Exists(fullpath))
+            {
+                return false;
+            }
+
             File.Delete(fullpath);
+            return true;
 
         }
-
-   }
+    }
 }
